proj1/src/requirements.txt
cobs==1.1.4
greenlet==1.1.2
importlib-metadata==4.8.2
pyzmq==22.3.0
SQLAlchemy==1.4.27
typing-extensions==4.0.0
zipp==3.6.0
